[reset] polish commit diff
